Add handling for base policy usage in Nessus results
The Nessus import handler should be updated to handle the unlikely, but
possible, chance that the base policy is used to start a scan on a vulnscan
instance. I ran into this in testing, and it's a small enough change that I
see no harm in rolling this into the codebase as a just-in-case.