Fix issue with calling pop with invalid index
The previous version used the pop() method to remove host dicts from the list
of hosts on cooldown. Unfortunately, if two or more hosts were removed in the
same loop, then the first call that references an index past the number of
removed hosts will result in failure. I changed it to use the remove() method
on the original list instead. There should be no duplicate hosts, and so in turn
there should only be one entry for a given host dict in the list of hosts on
cooldown.

I switched from list(self.__hosts_on_cooldown) to self.__hosts_on_cooldown[:] to
more clearly show that we just want a copy of the list, and are not trying to
convert a list into a list. The review on the original PR showed that it was not
particularly obvious at first skim.