Manually reconnect to hosts coming off of cooldown
We have been seeing the following exception when a host that has been on
cooldown is put back into rotation:
'NoneType' object has no attribute 'open_session'

Upon digging into it I realized that I naively assumed that Fabric would
reconnect to hosts when they were put back into rotation, but I believe it
simply checks that a connection exists and works with it without verifying
connectivity. This hopefully fixes the issue by forcing a reconnection to the
host when it is put back into rotation from cooldown.