Manually set the host_string environment value
When manually calling the HostCollectionCache.connect() method, it does not
pass the given host, nor does it set the host_string value, when making calls
to other methods/functions. This resulted in a
'NoneType' object has no attribute 'rsplit'
exception when trying to reconnect to hosts coming off of cooldown.