Update the VULNSCAN job to use a Nessus API configuration file
We added a configuration file for Nessus API access in
cisagov/cyhy_amis#542 which should be present on any host that is given
the VULNSCAN job (vulnscan.py) to run. This will allow this job to get
the configuration information that matches the configuration of the
specific runner that is given the job.