Add a default owner
The default owner becomes the owner of host documents created when a hostname (owned by a CyHy org other than the default owner) is resolved to an IP address that is not owned by any other orgs (owners) in the database.

If a request doc for the default owner (default: "CYHY") does not exist, it will be created at commander startup.