Check against multiple Nmap XML output versions
When processing the results from the `portscan.sh` job the XML output
version is checked for some logic. The previous version(s) of Nmap we
used produced XML output with a version of 1.04. The latest version we
are using (7.93) produces XML output with a version of 1.05. The check
against the XML output version is thus updated from checking for 1.04
to checking for either 1.04 or 1.05.