Modify __store_port_details() and __store_os_details() to handle hostnames more cleanly and consistently
When an IP whose HostDoc contains one or more hostnames is scanned, one host_scan or port_scan document is created for each hostname/owner combo.  This also means that we should not ever get any host_scan or port_scan documents that are owned by the CYHY default owner.