Process completed jobs using a thread pool
Instead of processing completed work retrieved from scanners it is
instead added to an appropriate queue and a pool of threads will
process the jobs in the queues. The commander's main loop will wait
for the work queue to empty before continuing on to mirror the
current behavior for the commander's loop.