Add a thread to output job queue load periodically
This thread will give some insight into how much work is on the two job
processing queues. It will output the "approximate size of the queue"
(per the Python documentation on the Queue.qsize() method) for the
successful and failed job queues every 10 seconds. The thread exits
alongside the job processing threads since they are related.

Co-authored-by: dav3r <david.redmin@gwe.cisa.dhs.gov>