Ensure threads report task completion
We ran into a bug where a job failed to process and as a result the
thread processing it never calls `Queue.task_done()`. The thread was
able to continue processing work, but without the call to
`Queue.task_done()`, the main commander loop hung on its `Queue.join()`
call. Thus we move this call out of the thread's `try`/`except` block
to ensure that `Queue.task_done()` is called no matter what. This also
follows best practices of minimizing the amount of code housed in a
`try`/`except` block.