Change the value used by the queue monitor
The `Queue.qsize()` method returns the "approximate size of the queue".
Under full load in testing this is usually a 0 because there are no
jobs unclaimed by a job processor thread. We instead output the value
of `Queue.unfinished_tasks` to hopefully better inform how many jobs
are still being processed.