Minimize `while` loop condition for job processing threads
Reduce the `while` loop to only check the value of
self.__is_processing_jobs. The previous logic was needed because the
value of self.__is_running could be change outside of the `do_work()`
method. With the new control variable it is only modified within the
`do_work()` method and is only set to `False` after the `do_work()`'s
main `while` loop has exited. As a result, there should be no work on
the job processing queues when this control variable is set to `False`.
Thus, it is safe to rely solely on this control variable for the job
processing threads' work loop.